requirements.txt
kafka-python
cassandra-driver
river
numpy
twilio
pyspark
py4j